+++ NEWS.md
@@ -21,3 +21,5 @@
 * Added `microeda_beta()` for initial Bray-Curtis beta diversity distances.
 * Added `as_beta_dist()` to extract beta diversity distances from
   `microeda_beta` objects.
+* Added `as_beta_matrix()` to extract beta diversity distances as a square
+  matrix.